+++ src/main/java/com/tysanclan/site/projectewok/beans/impl/RealmServiceImpl.java
@@ -429,19 +429,19 @@ class RealmServiceImpl implements
 	 */
 	@Override
 	public int getRequiredPetitionSignatures() {
-		int mcount = userService.countMembers();
+		Long mcount = userService.countMembers();
 
 		return petitionFormula(mcount);
 	}
 
-	static int petitionFormula(int mcount) {
+	static int petitionFormula(long mcount) {
 		if (mcount > 75) {
 			return 9;
 		} else if (mcount < 25) {
 			return 4;
 		}
 
-		return ((mcount - 25) / 10) + 4;
+		return (int) (((mcount - 25) / 10) + 4);
 	}
 
 	@OnFortuityEvent(MembershipTerminatedEvent.class)
