+++ src/main/java/com/tysanclan/site/projectewok/tasks/SenateElectionChecker.java
@@ -51,11 +51,10 @@ public class SenateElectionChecker extends PeriodicTask {
 	private UserDAO userDAO;
 
 	/**
-     * 
-     */
+	 * 
+	 */
 	public SenateElectionChecker() {
-		super("Senate election", "Democracy",
-		        ExecutionMode.DAILY);
+		super("Senate election", "Democracy", ExecutionMode.DAILY);
 	}
 
 	/**
@@ -74,36 +73,31 @@ public class SenateElectionChecker extends PeriodicTask {
 		filter2.setStartAfter(calendar.getTime());
 		filter2.addOrderBy("start", false);
 
-		SenateElection current = democracyService
-		        .getCurrentSenateElection();
+		SenateElection current = democracyService.getCurrentSenateElection();
 
 		if (senateElectionDAO.countAll() == 0
-		        || (senateElectionDAO.countByFilter(filter) > 0
-		                && senateElectionDAO
-		                        .countByFilter(filter2) == 0 && current == null)) {
+				|| (senateElectionDAO.countByFilter(filter) > 0
+						&& senateElectionDAO.countByFilter(filter2) == 0 && current == null)) {
 			democracyService.createSenateElection();
 		} else {
 			SenateElectionFilter filter3 = new SenateElectionFilter();
 			filter3.addOrderBy("start", false);
 			List<SenateElection> elections = senateElectionDAO
-			        .findByFilter(filter3);
+					.findByFilter(filter3);
 
 			UserFilter filter4 = new UserFilter();
 			filter4.addRank(Rank.SENATOR);
 
 			for (SenateElection election : elections) {
-				int seats = wrapperService
-				        .countElectionWinner(election);
+				int seats = wrapperService.countElectionWinner(election);
 
 				if (seats > 0) {
-					int senators = userDAO
-					        .countByFilter(filter4);
+					long senators = userDAO.countByFilter(filter4);
 
-					int fraction = (senators * 100) / seats;
+					int fraction = (int) ((senators * 100) / seats);
 
 					if (fraction < 40) {
-						democracyService
-						        .createSenateElection();
+						democracyService.createSenateElection();
 					}
 				}
 				break;
